chore: move getToken to useFetch
